1.1
-Criação de AdminUser
-Classe  de fazer reserva

+++ resources/hotel.py
@@ -1,6 +1,15 @@
-from flask_restful import Resource, reqparse
+from flask_restful import Resource, reqparse, inputs
 from models.hotel import HotelModel
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt, decode_token
+import json
+
+
+def is_admin():
+    tipo_usuario = get_jwt_identity()['tipo']
+    if "admin" in tipo_usuario:
+        return True
+    return False
+
 
 class Hoteis(Resource):
     def get(self):
@@ -12,12 +21,15 @@ class Hotel(Resource):
     argumentos.add_argument('estrelas', type=float, required=True, help="The fild 'estrelas' cannot be left blank")
     argumentos.add_argument('diaria')
     argumentos.add_argument('cidade')
+    # argumentos.add_argument('reserva')
+    # argumentos.add_argument('hospede')
 
-    def find_hotel(hotel_id):
-        for hotel in HotelModel.query.all():
-            if hotel['hotel_id'] == hotel_id:
-                return hotel
-        return None
+
+    # def find_hotel(hotel_id):
+    #     for hotel in HotelModel.query.all():
+    #         if hotel['hotel_id'] == hotel_id:
+    #             return hotel
+    #     return None
 
 #CRUD
     def get(self, hotel_id):
@@ -28,38 +40,68 @@ class Hotel(Resource):
 
     @jwt_required()
     def post(self, hotel_id):
-        if HotelModel.find_hotel(hotel_id):
-            return {"message": "Hotel id '{}' already exists".format(hotel_id)}, 400
-        dados = Hotel.argumentos.parse_args()
-        hotel = HotelModel(hotel_id,**dados)
-        try:
-            hotel.save_hotel()
-        except:
-            return {'message': 'An internal error ocurred trying to save hotel.'}, 500
-        return hotel.json()
+        if is_admin():
+            if HotelModel.find_hotel(hotel_id):
+                return {"message": "Hotel id '{}' already exists".format(hotel_id)}, 400
+            dados = Hotel.argumentos.parse_args()
+            hotel = HotelModel(hotel_id,**dados)
+            try:
+                hotel.save_hotel()
+            except:
+                return {'message': 'An internal error ocurred trying to save hotel.'}, 500
+            return hotel.json()
+        return {'message': 'Only administrators can create new hotels.'}, 401
 
     @jwt_required()
     def put(self, hotel_id):
-        dados = Hotel.argumentos.parse_args()
-        hotel_encontrado = HotelModel.find_hotel(hotel_id)
-        if hotel_encontrado:
-            hotel_encontrado.update_hotel(**dados)
-            hotel_encontrado.save_hotel()
-            return hotel_encontrado.json(),200
-        hotel = HotelModel(hotel_id, **dados)
-        try:
-            hotel.save_hotel()
-        except:
-            return {'message': 'An internal error ocurred trying to save hotel.'}, 500
-        return hotel.json(), 201
+        if is_admin():
+            dados = Hotel.argumentos.parse_args()
+            hotel_encontrado = HotelModel.find_hotel(hotel_id)
+            if hotel_encontrado:
+                hotel_encontrado.update_hotel(**dados)
+                hotel_encontrado.save_hotel()
+                return hotel_encontrado.json(),200
+            hotel = HotelModel(hotel_id, **dados)
+            try:
+                hotel.save_hotel()
+            except:
+                return {'message': 'An internal error ocurred trying to save hotel.'}, 500
+            return hotel.json(), 201
+        return {'message': 'Only administrators can edit hotels.'}, 401
 
     @jwt_required()
     def delete(self, hotel_id):
+        if is_admin():
+            hotel = HotelModel.find_hotel(hotel_id)
+            if hotel:
+                try:
+                    hotel.delete_hotel()
+                except:
+                    return {'message': 'An error ocurred trying to delet hotel.'}, 500
+                return {'message': 'Hotel deleted'}
+            return {'message': 'Hotel not found'}, 404
+        return {'message': 'Only administrators can create new hotels.'}, 401
+
+class HotelReserva(Resource):
+
+    @jwt_required()
+    def post(self, hotel_id):
+        hospede = get_jwt_identity()['login']
+        # print(f"Tipo: {get_jwt_identity()}")
+        argumentos = reqparse.RequestParser()
+        argumentos.add_argument('reserva', type=str, required=True, help="The fild 'reserva' cannot be left blank")
+        argumentos.add_argument('checkin', type=inputs.date , required=True, help="The fild 'reserva' cannot be left blank")
+        dados = argumentos.parse_args()
+
         hotel = HotelModel.find_hotel(hotel_id)
         if hotel:
             try:
-                hotel.delete_hotel()
+                hotel.reservar_quarto(dados['reserva'],hospede, dados['checkin'])
             except:
                 return {'message': 'An error ocurred trying to delet hotel.'}, 500
-            return {'message': 'Hotel deleted'}
+            return {'message': 'Hotel reservado com sucesso'}
         return {'message': 'Hotel not found'}, 404
+
+
+
+
